+++ app/services/tests/repoApi.test.js
@@ -1,8 +1,8 @@
 import MockAdapter from 'axios-mock-adapter';
 import { getApiClient } from '@utils/apiUtils';
-import { getSong } from '../repoApi';
+import { getSong } from '../musicApi';
 
-describe('RepoApi tests', () => {
+describe('musicApi tests', () => {
   const musicName = 'arijit';
   it('should make the api call to "/search?term=${musicName}"', async () => {
     const mock = new MockAdapter(getApiClient().axiosInstance);
